server/requirements.txt
django
django-cors-headers
djangorestframework-simplejwt
pillow
python-dotenv